add pid file check

+++ app/services/process_service.rb
@@ -7,8 +7,8 @@ class ProcessService
 
     pid = fork do
       # tracking if the op failed for the Process exit
+      pid_file = PidFile.new(:piddir => 'app/', :pidfile => "sugarboo.pid")
       success = true
-      PidFile.new(:piddir => 'app/', :pidfile => "sugarboo.pid")
       begin
         ActiveRecord::Base.establish_connection(config)
         # This is needed to re-initialize the random number generator after forking (if you want diff random numbers generated in the forks)
@@ -25,7 +25,7 @@ class ProcessService
 
       ensure
         ActiveRecord::Base.remove_connection
-        File.unlink('app/sugarboo.pid')
+        File.unlink('app/sugarboo.pid') if pid_file.pidfile_exists?
         Process.exit! success
       end
     end
